Some belt and braces defensive codeing for the File commands.

+++ Source/v2/Meadow.Cli/Commands/Current/File/FileDeleteCommand.cs
@@ -18,22 +18,28 @@ public class FileDeleteCommand : BaseDeviceCommand<FileDeleteCommand>
     {
         var connection = await GetCurrentConnection();
 
-        if (connection == null)
-        {
-            return;
-        }
-
         if (connection != null)
         {
             var fileList = await connection.GetFileList(false);
 
             if (MeadowFile == "all")
             {
-                foreach (var f in fileList)
+                if (fileList != null)
                 {
-                    var p = Path.GetFileName(f.Name);
-                    Logger?.LogInformation($"Deleting file '{p}' from device...");
-                    await connection.Device.DeleteFile(p, CancellationToken);
+                    foreach (var f in fileList)
+                    {
+                        if (connection.Device != null)
+                        {
+                            var p = Path.GetFileName(f.Name);
+
+                            Logger?.LogInformation($"Deleting file '{p}' from device...");
+                            await connection.Device.DeleteFile(p, CancellationToken);
+                        }
+                        else
+                        {
+                            Logger?.LogError($"No Device Found.");
+                        }
+                    }
                 }
             }
             else
@@ -46,16 +52,23 @@ public class FileDeleteCommand : BaseDeviceCommand<FileDeleteCommand>
                 }
                 else
                 {
-                    var wasRuntimeEnabled = await connection.Device.IsRuntimeEnabled(CancellationToken);
-
-                    if (wasRuntimeEnabled)
+                    if (connection.Device != null)
                     {
-                        Logger?.LogError($"The runtime must be disabled before doing any file management. Use 'meadow runtime disable' first.");
-                        return;
-                    }
+                        var wasRuntimeEnabled = await connection.Device.IsRuntimeEnabled(CancellationToken);
 
-                    Logger?.LogInformation($"Deleting file '{MeadowFile}' from device...");
-                    await connection.Device.DeleteFile(MeadowFile, CancellationToken);
+                        if (wasRuntimeEnabled)
+                        {
+                            Logger?.LogError($"The runtime must be disabled before doing any file management. Use 'meadow runtime disable' first.");
+                            return;
+                        }
+
+                        Logger?.LogInformation($"Deleting file '{MeadowFile}' from device...");
+                        await connection.Device.DeleteFile(MeadowFile, CancellationToken);
+                    }
+                    else
+                    {
+                        Logger?.LogError($"No Device Found.");
+                    }
                 }
             }
         }
